Refactor: Separate classes and tests in files
